requirements.txt
requests
python-dateutil
termcolor
emoji
colorama
fake-useragent
validators
pyfiglet
tqdm
tabulate
